refactor: move proposal creation schema to it's own file
